Separated Model into a number of smaller files to make it more manageable
