+++ lib/Hooks.ts
@@ -1,7 +1,9 @@
 /// <reference path="../_references.d.ts" />
 import instance = require('./Instance');
 
-export interface IHooks<TDocument, TInstance> {
+export = IHooks;
+
+interface IHooks<TDocument, TInstance> {
     creating? (document: TDocument);
     retrieved? (document: TDocument);
     ready? (instance: TInstance);
