+++ lib/Model.js
@@ -1,15 +1,16 @@
 /// <reference path="../_references.d.ts" />
 var MongoDB = require('mongodb');
-var Skmatc = require('skmatc');
 var Concoction = require('concoction');
 var Promise = require('bluebird');
-var util = require('util');
 var _ = require('lodash');
 var Iridium = require('./Core');
 var instance = require('./Instance');
 var Cursor = require('./Cursor');
 var idCacheController = require('./cacheControllers/IDDirector');
-var Omnom = require('./utils/Omnom');
+var ModelCache = require('./ModelCache');
+var ModelHelpers = require('./ModelHelpers');
+var ModelHandlers = require('./ModelHandlers');
+var ModelSpecificInstance = require('./ModelSpecificInstance');
 /**
  * An Iridium Model which represents a structured MongoDB collection
  * @class
@@ -306,8 +307,9 @@ var Model = (function () {
         });
         return Promise.resolve().then(function () {
             var queryOptions = { w: options.w, upsert: options.upsert, new: true };
-            if (options.upsert)
-                return _this._handlers.creatingDocuments(objects).map(function (object) {
+            if (options.upsert) {
+                var docs = _this._handlers.creatingDocuments(objects);
+                return docs.map(function (object) {
                     return new Promise(function (resolve, reject) {
                         _this.collection.findAndModify({ _id: object._id }, ["_id"], object, queryOptions, function (err, result) {
                             if (err)
@@ -316,6 +318,7 @@ var Model = (function () {
                         });
                     });
                 });
+            }
             else
                 return _this._handlers.creatingDocuments(objects).then(function (objects) {
                     return new Promise(function (resolve, reject) {
@@ -445,187 +448,5 @@ var Model = (function () {
     };
     return Model;
 })();
-exports.Model = Model;
-var ModelHelpers = (function () {
-    function ModelHelpers(model) {
-        this.model = model;
-        this._validator = new Skmatc(model.schema);
-        this.transform = new Concoction(model.options.transforms);
-    }
-    /**
-     * Validates a document to ensure that it matches the model's ISchema requirements
-     * @param {any} document The document to validate against the ISchema
-     * @returns {SkmatcCore.IResult} The result of the validation
-     */
-    ModelHelpers.prototype.validate = function (document) {
-        return this._validator.validate(document);
-    };
-    /**
-     * Creates a selector based on the document's unique _id field
-     * @param {object} document The document to render the unique selector for
-     * @returns {{_id: any}} A database selector which can be used to return only this document
-     */
-    ModelHelpers.prototype.selectOne = function (document) {
-        var testDoc = _.cloneDeep(document);
-        this.transform.reverse(testDoc);
-        return {
-            _id: testDoc._id
-        };
-    };
-    Object.defineProperty(ModelHelpers.prototype, "identifierField", {
-        /**
-         * Gets the field used in the ISchema to represent the document _id
-         */
-        get: function () {
-            var id = new String("");
-            var testDoc = {
-                _id: id
-            };
-            this.transform.apply(testDoc);
-            var idField = null;
-            for (var k in testDoc)
-                if (testDoc[k] === id) {
-                    idField = k;
-                    break;
-                }
-            return idField;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    /**
-     * Creates a selector based on the document's unique _id field in downstream format
-     * @param {any} id The downstream identifier to use when creating the selector
-     * @returns {object} A database selector which can be used to return only this document in downstream form
-     */
-    ModelHelpers.prototype.selectOneDownstream = function (id) {
-        if (_.isPlainObject(id))
-            return _.pick(id, this.identifierField);
-        else {
-            var conditions = {};
-            conditions[this.identifierField] = id;
-            return conditions;
-        }
-    };
-    /**
-     * Wraps the given document in an instance wrapper for use throughout the application
-     * @param {any} document The document to be wrapped as an instance
-     * @param {Boolean} isNew Whether the instance originated from the database or was created by the application
-     * @param {Boolean} isPartial Whether the document supplied contains all information present in the database
-     * @returns {any} An instance which wraps this document
-     */
-    ModelHelpers.prototype.wrapDocument = function (document, isNew, isPartial) {
-        return new this.model.Instance(document, isNew, isPartial);
-    };
-    /**
-     * Performs a diff operation between two documents and creates a MongoDB changes object to represent the differences
-     * @param {any} original The original document prior to changes being made
-     * @param {any} modified The document after changes were made
-     */
-    ModelHelpers.prototype.diff = function (original, modified) {
-        var omnom = new Omnom();
-        omnom.diff(original, modified);
-        return omnom.changes;
-    };
-    return ModelHelpers;
-})();
-exports.ModelHelpers = ModelHelpers;
-var ModelHandlers = (function () {
-    function ModelHandlers(model) {
-        this.model = model;
-    }
-    ModelHandlers.prototype.documentReceived = function (conditions, result, wrapper, options) {
-        var _this = this;
-        if (options === void 0) { options = {}; }
-        _.defaults(options, {
-            cache: true,
-            partial: false
-        });
-        return Promise.resolve(result).then(function (target) {
-            return Promise.resolve().then(function () {
-                // Trigger the received hook
-                if (_this.model.options.hooks.retrieved)
-                    return _this.model.options.hooks.retrieved(target);
-            }).then(function () {
-                // Cache the document if caching is enabled
-                if (_this.model.core.cache && options.cache && !options.fields) {
-                    var cacheDoc = _.cloneDeep(target);
-                    return _this.model.cache.set(cacheDoc);
-                }
-            }).then(function () {
-                // Transform the document
-                _this.model.helpers.transform.apply(target);
-                // Wrap the document and trigger the ready hook
-                var wrapped = wrapper(target, false, !!options.fields);
-                if (_this.model.options.hooks.ready && wrapped instanceof _this.model.Instance)
-                    return Promise.resolve(_this.model.options.hooks.ready(wrapped)).then(function () { return wrapped; });
-                return wrapped;
-            });
-        });
-    };
-    ModelHandlers.prototype.creatingDocuments = function (documents) {
-        var _this = this;
-        return Promise.all(documents.map(function (document) {
-            return Promise.resolve().then(function () {
-                if (_this.model.options.hooks.retrieved)
-                    return _this.model.options.hooks.creating(document);
-            }).then(function () {
-                var validation = _this.model.helpers.validate(document);
-                if (validation.failed)
-                    return Promise.reject(validation.error);
-                _this.model.helpers.transform.reverse(document);
-                return document;
-            });
-        }));
-    };
-    ModelHandlers.prototype.savingDocument = function (instance, changes) {
-        var _this = this;
-        return Promise.resolve().then(function () {
-            if (_this.model.options.hooks.saving)
-                return _this.model.options.hooks.saving(instance, changes);
-        }).then(function () { return instance; });
-    };
-    return ModelHandlers;
-})();
-exports.ModelHandlers = ModelHandlers;
-function ModelSpecificInstance(model, instanceType) {
-    var constructor = function (doc, isNew, isPartial) {
-        instanceType.call(this, model, doc, isNew, isPartial);
-    };
-    util.inherits(constructor, instanceType);
-    _.each(Object.keys(model.schema), function (property) {
-        Object.defineProperty(constructor.prototype, property, {
-            get: function () {
-                return this._modified[property];
-            },
-            set: function (value) {
-                this._modified[property] = value;
-            },
-            enumerable: true
-        });
-    });
-    return constructor;
-}
-exports.ModelSpecificInstance = ModelSpecificInstance;
-var ModelCache = (function () {
-    function ModelCache(model) {
-        this.model = model;
-    }
-    ModelCache.prototype.set = function (value) {
-        if (!this.model.cacheDirector || !this.model.cacheDirector.valid(value))
-            return Promise.resolve(value);
-        return this.model.core.cache.set(this.model.cacheDirector.buildKey(value), value);
-    };
-    ModelCache.prototype.get = function (conditions) {
-        if (!this.model.cacheDirector || !this.model.cacheDirector.validQuery(conditions))
-            return Promise.resolve(null);
-        return this.model.core.cache.get(this.model.cacheDirector.buildQueryKey(conditions));
-    };
-    ModelCache.prototype.clear = function (conditions) {
-        if (!this.model.cacheDirector || !this.model.cacheDirector.validQuery(conditions))
-            return Promise.resolve(false);
-        return this.model.core.cache.clear(this.model.cacheDirector.buildQueryKey(conditions));
-    };
-    return ModelCache;
-})();
+module.exports = Model;
 //# sourceMappingURL=Model.js.map
